refactor shanke app
