+++ src/shanke/shanke.hpp
@@ -1,0 +1,7 @@
+#pragma once
+#include "input.hpp"
+
+namespace shanke {
+    void shanke_input_cb(const joystick_event& ev);
+    void shanke_main();
+}
